backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.12
pydantic[dotenv]==1.10.12
python-multipart==0.0.6
python-dotenv==1.0.0

# Database & ORM
supabase==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# AI & LangChain
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.0
langchain-community==0.0.10
openai==1.6.1

# Redis for caching and sessions
redis==5.0.1
aioredis==2.0.1

# WebSocket support
websockets==12.0
socketio==5.10.0
python-socketio==5.10.0

# Queue system
celery==5.3.4
kombu==5.3.4

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Utilities
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
